+++ CHANGELOG.md
@@ -442,6 +442,7 @@
 - [^] 用 Popper.js 替代了 Tether。`Overlay` 组件的 `options` prop（以及多个组件的 `overlay-options` prop）现在都对应于 [Popper.js 的 `options` 参数](https://popper.js.org/popper-documentation.html#new_Popper_new)。`Overlay` 的 `position` prop 在之前的版本就已经兼容了 Popper.js 的语法，所以如果之前就以 Popper.js 风格进行描述，则无需调整。见 [#574](https://github.com/ecomfe/veui/pull/574)。
 - [^] `Input` 组件的 `before`/`after` slot 重命名为 `prepend`/`append`，新的 `before`/`after` 位置在输入框外部。并且新增了 `before-label`/`after-label` 两个 slot。见 [#502](https://github.com/ecomfe/veui/issues/502)。
 - [^] `NumberInput` 组件的 `before`/`after` slot 重命名为 `prepend`/`append`。见 [#502](https://github.com/ecomfe/veui/issues/502)。
+- [-] 移除 `Alert` 组件的 `close-label` 属性。
 - [^] `Pagination` 组件根据新的设计不再支持总数显示。
 - [-] 移除已废弃的 `Pager` 组件，请使用 `Pagination` 组件代替。
 - [-] 移除 `Progress` 组件已废弃的 `precision` prop，请使用 `decimal-place` prop 代替。
@@ -469,6 +470,7 @@
 - [+] 为 `Overlay`/`Dialog` 组件新增了 `inline` prop，允许内联在内容中进行渲染。
 - [+] 为 `Overlay` 组件新增了 `match-width` prop，允许对于相对指定元素定位时，以目标元素宽度为浮层最小宽度。
 - [+] 为 `Progress` 组件新增了 `append` slot。
+- [+] 为 `Alert` 组件新增了 `title`、`extra` slot。
 
 ### 🐞 问题修复